Add Todoable::Resources::Item and Todoable::Resources::List into its own file
There are some consumers of this classes using it directly; there's no
need to have it define in the same file as the collection of list
resource.
