+++ lib/todoable/resources.rb
@@ -9,4 +9,6 @@ module Todoable
 end
 
 require_relative "./resources/token"
+require_relative "./resources/item"
+require_relative "./resources/list"
 require_relative "./resources/lists"
